get units tests working
