+++ requirements.txt
@@ -11,4 +11,5 @@ pylint==3.0.2
 black==23.11.0
 pytest==7.4.3
 pytest-cov==4.1.0
+pre-commit==3.6.0
 git+https://github.com/callummcdougall/eindex.git
