+++ sae_training/lm_runner.py
@@ -30,11 +30,8 @@ def language_model_sae_runner(cfg):
         model, sparse_autoencoder, activations_loader,
         n_checkpoints=cfg.n_checkpoints,
         batch_size = cfg.train_batch_size,
-        feature_sampling_method = cfg.feature_sampling_method,
         feature_sampling_window = cfg.feature_sampling_window,
-        feature_reinit_scale = cfg.feature_reinit_scale,
         dead_feature_threshold = cfg.dead_feature_threshold,
-        dead_feature_window=cfg.dead_feature_window,
         use_wandb = cfg.log_to_wandb,
         wandb_log_frequency = cfg.wandb_log_frequency
     )
